requirements/requirements.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.5

# Configuration
django-environ==0.4.1
whitenoise==3.3.0
django-debug-toolbar==1.9.1

# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6.1

# Images
Pillow==4.0.0
django-imagekit==4.0.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.30.0


# Python-PostgreSQL Database Adapter



# Time zones support
pytz==2016.10

# Redis support
django-redis==4.7.0
redis>=2.10.5

# Django extensions
django-extensions==1.7.6


# Your custom requirements go here

django-mptt==0.8.7
django-digg-paginator==0.1.3

# Celery
celery==4.0.2
amqp==2.1.4
kombu==4.0.2
flower==0.9.1

# Django REST framework
djangorestframework==3.5.4
markdown==2.6.8






